third_party/blink/web_tests/external/wpt/digital-credentials/get.tentative.https-expected.txt

This is a testharness.js-based test.
[FAIL] navigator.credentials.get() promise is rejected if abort controller is aborted after call to get() in cross-origin iframe.
  assert_equals: expected "AbortError" but got "NetworkError"
[FAIL] Mediation is required to get a DigitalCredential.
  assert_unreached: Should have rejected: undefined Reached unreachable code
Harness: the test ran to completion.
